extract out code editor logic
