+++ codepen-clone/.eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/index.tsx":"1","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/App.tsx":"2","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/components/code-editor.tsx":"3","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/components/preview.tsx":"4","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/bundler/index.ts":"5","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/bundler/plugins/unpkg-path-plugin.ts":"6","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/bundler/plugins/fetch-plugin.ts":"7"},{"size":249,"mtime":1623241596645,"results":"8","hashOfConfig":"9"},{"size":714,"mtime":1627484254931,"results":"10","hashOfConfig":"9"},{"size":1742,"mtime":1623237859200,"results":"11","hashOfConfig":"9"},{"size":998,"mtime":1626437302041,"results":"12","hashOfConfig":"9"},{"size":779,"mtime":1627485762927,"results":"13","hashOfConfig":"9"},{"size":826,"mtime":1624468967693,"results":"14","hashOfConfig":"9"},{"size":2099,"mtime":1625667073896,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},"ssvdrz",{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/index.tsx",[],["31","32"],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/App.tsx",[],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/components/code-editor.tsx",[],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/components/preview.tsx",[],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/bundler/index.ts",[],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/bundler/plugins/unpkg-path-plugin.ts",[],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/bundler/plugins/fetch-plugin.ts",[],{"ruleId":"33","replacedBy":"34"},{"ruleId":"35","replacedBy":"36"},"no-native-reassign",["37"],"no-negated-in-lhs",["38"],"no-global-assign","no-unsafe-negation"]
+[{"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/index.tsx":"1","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/App.tsx":"2","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/components/code-editor.tsx":"3","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/components/preview.tsx":"4","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/bundler/index.ts":"5","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/bundler/plugins/unpkg-path-plugin.ts":"6","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/bundler/plugins/fetch-plugin.ts":"7","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/components/code-cell.tsx":"8"},{"size":249,"mtime":1623241596645,"results":"9","hashOfConfig":"10"},{"size":365,"mtime":1627486082988,"results":"11","hashOfConfig":"10"},{"size":1742,"mtime":1623237859200,"results":"12","hashOfConfig":"10"},{"size":998,"mtime":1626437302041,"results":"13","hashOfConfig":"10"},{"size":779,"mtime":1627485762927,"results":"14","hashOfConfig":"10"},{"size":826,"mtime":1624468967693,"results":"15","hashOfConfig":"10"},{"size":2099,"mtime":1625667073896,"results":"16","hashOfConfig":"10"},{"size":678,"mtime":1627485970860,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"ssvdrz",{"filePath":"21","messages":"22","errorCount":0,"warningCount":4,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/index.tsx",[],["35","36"],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/App.tsx",["37","38","39","40"],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/components/code-editor.tsx",[],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/components/preview.tsx",[],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/bundler/index.ts",[],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/bundler/plugins/unpkg-path-plugin.ts",[],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/bundler/plugins/fetch-plugin.ts",[],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/components/code-cell.tsx",[],{"ruleId":"41","replacedBy":"42"},{"ruleId":"43","replacedBy":"44"},{"ruleId":"45","severity":1,"message":"46","line":2,"column":10,"nodeType":"47","messageId":"48","endLine":2,"endColumn":18},{"ruleId":"45","severity":1,"message":"49","line":3,"column":8,"nodeType":"47","messageId":"48","endLine":3,"endColumn":18},{"ruleId":"45","severity":1,"message":"50","line":4,"column":8,"nodeType":"47","messageId":"48","endLine":4,"endColumn":15},{"ruleId":"45","severity":1,"message":"51","line":5,"column":8,"nodeType":"47","messageId":"48","endLine":5,"endColumn":14},"no-native-reassign",["52"],"no-negated-in-lhs",["53"],"@typescript-eslint/no-unused-vars","'useState' is defined but never used.","Identifier","unusedVar","'CodeEditor' is defined but never used.","'Preview' is defined but never used.","'bundle' is defined but never used.","no-global-assign","no-unsafe-negation"]
